feat: add retry policy for secret client
