+++ internal/secret/options.go
@@ -1,0 +1,28 @@
+package secret
+
+import (
+	"time"
+
+	"github.com/KarlGW/azcfg/internal/httpr"
+)
+
+// WithConcurrency sets the concurrency for secret retrieval.
+func WithConcurrency(n int) ClientOption {
+	return func(c *Client) {
+		c.concurrency = n
+	}
+}
+
+// WithTimeout sets timeout for secret retreival.
+func WithTimeout(d time.Duration) ClientOption {
+	return func(c *Client) {
+		c.timeout = d
+	}
+}
+
+// WithRetryPolicy sets the retry policy for secret retrieval.
+func WithRetryPolicy(r httpr.RetryPolicy) ClientOption {
+	return func(c *Client) {
+		c.retryPolicy = r
+	}
+}
